🏷️ (dev): Migrate from global types
