--- src/script/lib/globals.d.ts
@@ -1,106 +1,0 @@
-interface BackgroundData {
-  disabledTab?: boolean;
-}
-
-interface BackgroundStorage {
-  tabs?: {
-    [tabId: number]: TabStorageOptions;
-  };
-}
-
-interface ConfirmModalSettings {
-  backup?: boolean;
-  content?: string;
-  title?: string;
-  titleClass?: string;
-}
-
-interface DomainCfg {
-  adv?: boolean;
-  deep?: boolean;
-  disabled?: boolean;
-  enabled?: boolean;
-  framesOff?: boolean;
-  framesOn?: boolean;
-  wordlist?: number;
-}
-
-interface Message {
-  advanced?: boolean;
-  backgroundData?: boolean;
-  counter?: number;
-  deep?: boolean;
-  destination: string;
-  disabled?: boolean;
-  enableTab?: boolean;
-  forceUpdate?: boolean;
-  getStatus?: boolean;
-  globalVariable?: string;
-  iframe?: boolean;
-  popup?: boolean;
-  source: string;
-  status?: number;
-  summary?: Summary;
-  tabId?: number;
-  updateContextMenus?: boolean;
-  urlUpdate?: string;
-}
-
-interface Migration {
-  async?: boolean;
-  name: string;
-  runOnImport: boolean;
-  version: string;
-}
-
-interface ReplaceTextResult {
-  filtered: string;
-  modified: boolean;
-  original: string;
-}
-
-interface Statistics {
-  startedAt?: number;
-  words: WordStatistics;
-}
-
-interface Summary {
-  [word: string]: {
-    filtered: string;
-    count: number;
-  };
-}
-
-interface TabStorageOptions {
-  counters?: { number?: number };
-  disabled?: boolean;
-  disabledOnce?: number; // NOT_SET: 0, DISABLED: 1, WILL_DISABLE: 2
-  id?: number;
-  registeredAt?: number;
-  status?: number;
-}
-
-interface Version {
-  major: number;
-  minor: number;
-  patch: number;
-}
-
-interface WordOptions {
-  _filterMethod?: number; // This should not be stored in the config. Only there for new Word
-  case?: number;
-  lists?: number[];
-  matchMethod: number;
-  repeat: number;
-  separators?: number;
-  sub: string;
-}
-
-interface WordStatistic {
-  text: number;
-  total?: number;
-}
-
-interface WordStatistics {
-  [word: string]: WordStatistic;
-}
